refactor: structured codebase
