+++ src/App.js
@@ -1,9 +1,10 @@
 // import logo from './logo.svg';
 // import "./App.css";
 import "./scss/app.scss";
-import { Header } from "./header";
-import { Body } from "./body";
-import { Footer } from "./footer";
+import { Header } from "./components/header";
+import { Body } from "./components/body";
+import { Footer } from "./components/footer";
+
 function App() {
 	return (
 		<div class="container">
