--- src/index.css
@@ -1,105 +1,0 @@
-@import url("https://fonts.googleapis.com/icon?family=Material+Icons");
-.container {
-  height: 100% !important;
-  width: 100% !important;
-  display: -ms-grid;
-  display: grid;
-  -ms-grid-rows: auto 1fr auto;
-      grid-template-rows: auto 1fr auto;
-}
-
-.header {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: row;
-          flex-direction: row;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  background-color: #eb5d05;
-  color: #0a0706;
-}
-
-.header .heading {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-flex: 1;
-      -ms-flex: 1;
-          flex: 1;
-  font-size: 5vh;
-  margin-left: 2vh;
-}
-
-.header .container-list ul {
-  list-style-type: none;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  text-align: center;
-  -webkit-box-pack: end;
-      -ms-flex-pack: end;
-          justify-content: flex-end;
-  font-size: 3vh;
-}
-
-.header .container-list ul li {
-  margin: 0 2vh 0 2vh;
-}
-
-.content {
-  font-size: 2vh;
-  color: yellow;
-  background-color: blue;
-}
-
-.footer {
-  font-size: 2vh;
-  background-color: red;
-}
-
-.header {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: row;
-          flex-direction: row;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  background-color: #eb5d05;
-  color: #0a0706;
-}
-
-.header .heading {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-flex: 1;
-      -ms-flex: 1;
-          flex: 1;
-  font-size: 5vh;
-  margin-left: 2vh;
-}
-
-.header .container-list ul {
-  list-style-type: none;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  text-align: center;
-  -webkit-box-pack: end;
-      -ms-flex-pack: end;
-          justify-content: flex-end;
-  font-size: 3vh;
-}
-
-.header .container-list ul li {
-  margin: 0 2vh 0 2vh;
-}
-/*# sourceMappingURL=index.css.map */
