+++ src/scss/app.scss
@@ -1,3 +1,8 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
 .container{
   height: 100% !important;
   width: 100% !important;
